+++ src/main/java/com/team5/catdogeats/batch/mapper/SettlementChunkMapper.java
@@ -1,29 +1,22 @@
 package com.team5.catdogeats.batch.mapper;
 
-import com.team5.catdogeats.batch.dto.SettlementBatchItem;
 import org.apache.ibatis.annotations.*;
 
-import java.util.List;
-
-
+/**
+ * 정산 배치용 MyBatis Mapper (정리된 버전)
+ * Admin Controller에서 현황 조회용으로만 사용
+ */
 @Mapper
 public interface SettlementChunkMapper {
 
     /**
-     * 정산 생성 대상 주문아이템 조회
-     * 성능을 위해 최근 10일로 제한
+     * 정산 생성 대상 전체 건수 (Admin Controller용)
      */
     @Select("""
-        SELECT 
-            s.seller_id,
-            oi.id as order_item_id,
-            o.order_number,
-            p.title as product_title,
-            oi.price as item_price
+        SELECT COUNT(*)
         FROM shipments s
-        INNER JOIN orders o ON s.order_id = o.id 
+        INNER JOIN orders o ON s.order_id = o.id
         INNER JOIN order_items oi ON o.id = oi.order_id
-        INNER JOIN products p ON oi.product_id = p.id
         WHERE s.delivered_at IS NOT NULL
           AND s.delivered_at <= CURRENT_DATE - INTERVAL '7 days'
           AND s.delivered_at >= CURRENT_DATE - INTERVAL '14 days'
@@ -33,115 +26,11 @@ public interface SettlementChunkMapper {
               SELECT 1 FROM settlements st 
               WHERE st.order_item_id = oi.id
           )
-        ORDER BY s.delivered_at ASC, oi.id ASC
-        LIMIT #{limit} OFFSET #{offset}
-        """)
-    @Results({
-            @Result(property = "sellerId", column = "seller_id"),
-            @Result(property = "orderItemId", column = "order_item_id"),
-            @Result(property = "orderNumber", column = "order_number"),
-            @Result(property = "productTitle", column = "product_title"),
-            @Result(property = "itemPrice", column = "item_price")
-    })
-    List<SettlementBatchItem> findUnsettledItems(@Param("offset") int offset,
-                                                 @Param("limit") int limit);
-
-    /**
-     * 정산 완료 대상 조회
-     * 오래된 IN_PROGRESS 정산은 처리할 필요 없으므로 2개월로 제한
-     */
-    @Select("""
-        SELECT 
-            st.id as settlement_id,
-            st.seller_id,
-            o.order_number,
-            p.title as product_title,
-            st.settlement_amount
-        FROM settlements st
-        INNER JOIN order_items oi ON st.order_item_id = oi.id
-        INNER JOIN orders o ON oi.order_id = o.id
-        INNER JOIN products p ON oi.product_id = p.id
-        WHERE st.settlement_status = 'IN_PROGRESS'
-          AND st.created_at >= CURRENT_DATE - INTERVAL '60 days'
-        ORDER BY st.created_at ASC, st.id ASC
-        LIMIT #{limit} OFFSET #{offset}
-        """)
-    @Results({
-            @Result(property = "settlementId", column = "settlement_id"),
-            @Result(property = "sellerId", column = "seller_id"),
-            @Result(property = "orderNumber", column = "order_number"),
-            @Result(property = "productTitle", column = "product_title"),
-            @Result(property = "settlementAmount", column = "settlement_amount")
-    })
-    List<SettlementBatchItem> findInProgressSettlements(@Param("offset") int offset,
-                                                        @Param("limit") int limit);
-
-    /**
-     * 정산 데이터 생성
-     */
-    @Insert("""
-        INSERT INTO settlements (
-            id,
-            seller_id,
-            order_item_id,
-            item_price,
-            commission_rate,
-            commission_amount,
-            settlement_amount,
-            settlement_status,
-            created_at,
-            updated_at
-        ) VALUES (
-            gen_random_uuid(),
-            #{sellerId},
-            #{orderItemId},
-            #{itemPrice},
-            #{commissionRate},
-            #{commissionAmount},
-            #{settlementAmount},
-            'IN_PROGRESS',
-            NOW(),
-            NOW()
-        )
-        """)
-    void insertSettlement(SettlementBatchItem item);
-
-    /**
-     * 정산 상태를 COMPLETED로 변경
-     */
-    @Update("""
-        UPDATE settlements 
-        SET settlement_status = 'COMPLETED',
-            settled_at = NOW(),
-            updated_at = NOW()
-        WHERE id = #{settlementId}
-          AND settlement_status = 'IN_PROGRESS'
-        """)
-    int updateToCompleted(@Param("settlementId") String settlementId);
-
-    /**
-     * 정산 생성 대상 전체 건수
-     */
-    @Select("""
-        SELECT COUNT(*)
-        FROM shipments s
-        INNER JOIN orders o ON s.order_id = o.id
-        INNER JOIN order_items oi ON o.id = oi.order_id
-        WHERE s.delivered_at IS NOT NULL
-         AND s.delivered_at <= CURRENT_DATE - INTERVAL '7 days'
-        AND s.delivered_at >= CURRENT_DATE - INTERVAL '14 days'
-          AND o.order_status = 'DELIVERED'
-          AND o.is_hidden = false
-          AND NOT EXISTS (
-              SELECT 1 FROM settlements st 
-              WHERE st.order_item_id = oi.id
-          )
         """)
     int countUnsettledItems();
 
-
     /**
-     * 정산 완료 대상 전체 건수
+     * 정산 완료 대상 전체 건수 (Admin Controller용)
      */
     @Select("""
         SELECT COUNT(*)
